Refactor help message into Ui
