+++ src/main/java/seedu/duke/commands/CommandHandler.java
@@ -3,7 +3,7 @@ package seedu.duke.commands;
 import seedu.duke.exceptions.DukeError;
 import seedu.duke.exercisegenerator.GenerateExercise;
 import seedu.duke.ui.Ui;
-import seedu.duke.util.OuputText;
+import seedu.duke.ui.PrintHelpMessage;
 
 public class CommandHandler {
 
@@ -22,8 +22,7 @@ public class CommandHandler {
                 System.exit(0);
                 break;
             case "help":
-                OuputText outputText = new OuputText();
-                outputText.showAvailableCommands();
+                ui.
                 break;
             default:
                 System.out.println("Unknown Command");
